refactor code for BoardFavorites
