+++ imports/services/index.ts
@@ -1,3 +1,4 @@
 export * from './ServiceManagerServer'
 export * from './AccountService'
 export * from './ExcalidrawSyncService'
+export * from './BoardService'
