+++ imports/subscriptions/index.ts
@@ -1,2 +1,3 @@
 export * from './BoardsSubscriber'
 export * from './SubscribersManager'
+export * from './BoardFavoritesSubscriber'
